Integer Input

+++ cards/forms.py
@@ -8,11 +8,11 @@ BANK_CHOICES   = [('chase', 'Chase'), ('citi', 'Citibank'), ('amex', 'American E
 
 class CreditForm(forms.Form):
 
-    groceries = forms.DecimalField(label='Groceries', min_value=0)
-    dining = forms.DecimalField(label='Dining out', min_value=0)
-    gas = forms.DecimalField(label='Gas', min_value=0)
-    travels = forms.DecimalField(label='Travels', min_value=0)
-    etc = forms.DecimalField(label='Everything Else', min_value=0)
+    groceries = forms.IntegerField(label='Groceries', min_value=0)
+    dining = forms.IntegerField(label='Dining out', min_value=0)
+    gas = forms.IntegerField(label='Gas', min_value=0)
+    travels = forms.IntegerField(label='Travels', min_value=0)
+    etc = forms.IntegerField(label='Everything Else', min_value=0)
     credit_score = forms.CharField(label='What is your credit score?', widget= forms.Select(choices= CREDIT_CHOICES))
     annual= forms.ChoiceField(label='Are you only interested in cards with no annual fee?', widget= forms.RadioSelect, choices= ANNUAL_CHOICES, initial= 'no')
     banks = forms.MultipleChoiceField(label= 'Show cards from the following banks:', widget= forms.CheckboxSelectMultiple, choices= BANK_CHOICES, initial= 'all')
